Feedback on grid

+++ src/systems/selection-system.h
@@ -15,11 +15,11 @@ public:
 
 private:
     scomp::Face colorToFace(unsigned char color) const;
-    scomp::Face normalToFace(const glm::vec3& normal) const;
+    scomp::Face normalToFace(const glm::ivec3& normal) const;
 
 private:
     Context& m_ctx;
     SingletonComponents& m_scomps;
-    std::array<glm::vec3, 6> m_planeNormals;
+    std::array<glm::ivec3, 6> m_planeNormals;
     std::array<glm::vec3, 6> m_planePositions;
 };
